pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cable_routing"
version = "0.0.1"
description = "YuMi Externally Guided Motion (EGM) wrapper"
authors = [
    {name = "osheraz", email = "osheraz@berkeley.edu"},
]
license = {text = "MIT Software License"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "viser",
    "tyro",
    "hydra-core",
    "opencv_python",
    "matplotlib",
    "omegaconf"
]

[tool.setuptools.packages.find]
exclude = ["catkin_ws*"]
